fix some C4244 warnings
git-svn-id: https://mpc-hc.svn.sourceforge.net/svnroot/mpc-hc/trunk@4649 10f7b99b-c216-0410-bff0-8a66a9350fd8

+++ src/apps/mplayerc/PPageAudioSwitcher.cpp
@@ -327,7 +327,7 @@ void CPPageAudioSwitcher::OnHScroll(UINT nSBCode, UINT nPos, CScrollBar* pScroll
 {
 	if (*pScrollBar == m_AudioBoostCtrl) {
 		UpdateData();
-		((CMainFrame*)GetParentFrame())->SetVolumeBoost(m_AudioBoostPos/10.0); // nice shortcut...
+		((CMainFrame*)GetParentFrame())->SetVolumeBoost((float)m_AudioBoostPos/10); // nice shortcut...
 	}
 
 	SetModified();
